+++ app/src/main/java/com/ayush/weatherapp/mapper/ForecastDTOtoRealmMapper.java
@@ -7,12 +7,12 @@ import com.ayush.weatherapp.realm.model.forecast.DailyForecast;
 import com.ayush.weatherapp.realm.model.forecast.Forecast;
 import com.ayush.weatherapp.realm.model.forecast.HourlyData;
 import com.ayush.weatherapp.realm.model.forecast.HourlyForecast;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.CurrentForecastDTO;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.DailyDataDTO;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.DailyForecastDTO;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.ForecastDTO;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.HourlyDataDTO;
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.HourlyForecastDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.CurrentForecastDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.DailyDataDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.DailyForecastDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.ForecastDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.HourlyDataDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.HourlyForecastDTO;
 import com.ayush.weatherapp.utils.DateUtils;
 import java.util.ArrayList;
 import java.util.List;
