+++ app/src/main/java/com/ayush/weatherapp/retrofit/weatherApi/WeatherAPIInterface.java
@@ -1,8 +1,7 @@
 package com.ayush.weatherapp.retrofit.weatherApi;
 
-import com.ayush.weatherapp.retrofit.weatherApi.pojo.ForecastDTO;
+import com.ayush.weatherapp.retrofit.weatherApi.model.ForecastDTO;
 import io.reactivex.Observable;
-import io.reactivex.Single;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
