+++ app/src/main/java/com/ayush/weatherapp/retrofit/weatherApi/model/DailyForecastDTO.java
@@ -1,12 +1,12 @@
-package com.ayush.weatherapp.retrofit.weatherApi.pojo;
+package com.ayush.weatherapp.retrofit.weatherApi.model;
 
 import com.google.gson.annotations.SerializedName;
 import java.util.List;
 
-public class HourlyForecastDTO {
+public class DailyForecastDTO {
   @SerializedName("summary") private String summary;
   @SerializedName("icon") private String icon;
-  @SerializedName("data") private List<HourlyDataDTO> hourlyDataDTOList;
+  @SerializedName("data") private List<DailyDataDTO> dailyDataDTOList;
 
   public String getSummary() {
     return summary;
@@ -16,7 +16,7 @@ public class HourlyForecastDTO {
     return icon;
   }
 
-  public List<HourlyDataDTO> getHourlyDataDTOList() {
-    return hourlyDataDTOList;
+  public List<DailyDataDTO> getDailyDataDTOList() {
+    return dailyDataDTOList;
   }
 }
